[u] Split TDR plugin into tdr_hca and ABC (#4224)
