+++ deployments/prod/environment.py
@@ -503,7 +503,7 @@ def env() -> Mapping[str, Optional[str]]:
             f'{catalog}{suffix}': dict(atlas=atlas,
                                        internal=internal,
                                        plugins=dict(metadata=dict(name='hca'),
-                                                    repository=dict(name='tdr')),
+                                                    repository=dict(name='tdr_hca')),
                                        sources=list(filter(None, sources.values())))
             for atlas, catalog, sources in [
                 ('hca', 'dcp17', dcp17_sources),
