+++ scripts/generate_openapi_document.py
@@ -26,7 +26,7 @@ def main():
                                atlas='hca',
                                internal=False,
                                plugins=dict(metadata=config.Catalog.Plugin(name='hca'),
-                                            repository=config.Catalog.Plugin(name='tdr')),
+                                            repository=config.Catalog.Plugin(name='tdr_hca')),
                                sources=set())
     }
