+++ test/service/test_response.py
@@ -3723,7 +3723,7 @@ class TestTDRIndexer(WebServiceTestCase, TDRPluginTestCase):
                                         atlas='hca',
                                         internal=False,
                                         plugins=dict(metadata=config.Catalog.Plugin(name='hca'),
-                                                     repository=config.Catalog.Plugin(name='tdr')),
+                                                     repository=config.Catalog.Plugin(name='tdr_hca')),
                                         sources=set())
         }
