+++ test/test_doctests.py
@@ -32,7 +32,7 @@ import azul.openapi.responses
 import azul.openapi.schema
 import azul.plugins.metadata.hca.indexer.transform
 import azul.plugins.metadata.hca.service.contributor_matrices
-import azul.plugins.repository.tdr
+import azul.plugins.repository.tdr_hca
 import azul.service.drs_controller
 import azul.service.manifest_service
 import azul.service.repository_controller
@@ -79,7 +79,7 @@ def load_tests(_loader, tests, _ignore):
         azul.openapi.responses,
         azul.openapi.schema,
         azul.plugins.metadata.hca.service.contributor_matrices,
-        azul.plugins.repository.tdr,
+        azul.plugins.repository.tdr_hca,
         azul.plugins.metadata.hca.indexer.transform,
         azul.service.drs_controller,
         azul.service.manifest_service,
